refactor (wip): create editor state
