+++ frontend/level_editor/main.py
@@ -4,65 +4,11 @@ import pygame
 import pygame_gui
 from pygame_gui.elements import UIPanel
 import os
-from typing import List, Optional
+from typing import Optional
 import json
 
-
-class Vec2:
-    def __init__(self, x: int, y: int):
-        self.x: int = x
-        self.y: int = y
-
-
-class Station:
-    def __init__(self, name: str, asset_file: str, pos: Vec2):
-        self.name = name
-        self.asset_file = asset_file
-        self.pos = pos
-
-
-class PredicateDict(dict):
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-
-    def __getitem__(self, key: set):
-        frozenkey: frozenset = frozenset(key) if isinstance(key, set) else key
-        return super().__getitem__(frozenkey)
-
-    def __setitem__(self, key: set, value):
-        frozenkey: frozenset = frozenset(key) if isinstance(key, set) else key
-        return super().__setitem__(frozenkey, value)
-
-    def __contains__(self, key):
-        if isinstance(key, set):
-            key = frozenset(key)
-        return super().__contains__(key)
-
-
-class Item:
-    def __init__(self, name: str, state_map: list[tuple[list[str], str]]):
-        self.name = name
-        frozen_state_map = {frozenset(k): v for k, v in state_map}
-        self.state_map: PredicateDict = PredicateDict(frozen_state_map)
-
-
-class ItemInstance:
-    def __init__(self, source_item: Item, predicates: set[str], pos: Vec2):
-        self.source_item = source_item
-        self.state = source_item.state_map[frozenset(predicates)]
-        self.predicates = predicates
-        self.pos = pos
-
-    def get_asset(self, predicates: List[str]) -> Optional[str]:
-        """
-        This function takes a list of string predicates and returns an option(str)
-        representing an some/none filepath.
-        """
-        try:
-            return self.source_item.state_map[frozenset(predicates)]
-        except KeyError:
-            return None
-
+from declarations import Vec2, Item, Station, ItemInstance
+from level_state import LevelState, NoStationAtLocationError
 
 root = tk.Tk()
 root.withdraw()
@@ -86,7 +32,6 @@ config_file_path = (
     if not dev_mode
     else "renderer/configuration/robotouille_config.json"
 )
-
 
 if not config_file_path:
     print("No config file selected. Exiting.")
@@ -113,121 +58,6 @@ for item_name, item_data in item_entities.items():
         ):
             state_map.append((state_info["predicates"], state_info["asset"]))
     items[item_name] = Item(item_name, state_map)
-
-print(items)
-
-
-fried_chicken = Item(
-    "chicken",
-    [
-        ([], "chicken.png"),
-        (["isfried"], "friedchicken.png"),
-        (["iscooked"], "cookedchicken.png"),
-    ],
-)
-
-
-class LevelState:
-    def __init__(self, width: int, height: int):
-        self.width: int = width
-        self.height: int = height
-        self._stations: List[Station] = []
-        self._items: List[ItemInstance] = []
-
-    def get_station_at(self, pos: Vec2) -> Optional[Station]:
-        for station in self._stations:
-            if station.pos.x == pos.x and station.pos.y == pos.y:
-                return station
-        return None
-
-    def get_item_at(self, pos: Vec2) -> Optional[ItemInstance]:
-        for item in self._items:
-            if item.pos.x == pos.x and item.pos.y == pos.y:
-                return item
-        return None
-
-    def get_all_stations(self) -> List[Station]:
-        return self._stations
-
-    def get_all_items(self) -> List[Item]:
-        return self._items
-
-    def put_station_at(self, station: Station):
-        pos = station.pos
-        existing_station = self.get_station_at(pos)
-        if existing_station is None:
-            self._stations.append(station)
-        else:
-            self._stations.remove(existing_station)
-            self._stations.append(station)
-
-    def put_item_at(self, item: ItemInstance):
-        pos = item.pos
-        existing_station = self.get_station_at(pos)
-        if existing_station is None:
-            raise NoStationAtLocationError(
-                "Cannot place item at location without a station"
-            )
-
-        existing_item = self.get_item_at(pos)
-        if existing_item is None:
-            self._items.append(item)
-        else:
-            self._items.remove(existing_item)
-            self._items.append(item)
-
-    def serialize(self) -> dict:
-        stations_json = []
-        for station in self.get_all_stations():
-            stations_json.append(
-                {
-                    "name": station.name,
-                    "x": station.pos.x,
-                    "y": self.height - 1 - station.pos.y,
-                }
-            )
-
-        items_json = []
-        for item in self.get_all_items():
-            items_json.append(
-                {
-                    "name": item.source_item.name,
-                    "x": item.pos.x,
-                    "y": self.height - 1 - item.pos.y,
-                    "stack-level": 0,  # This is hardcoded for now
-                    "predicates": list(item.predicates),
-                }
-            )
-
-        level_json = {
-            "version": "1.0.0",
-            "width": self.width,
-            "height": self.height,
-            "config": {
-                "num_cuts": {"lettuce": 3, "default": 3},
-                "cook_time": {"patty": 3, "default": 3},
-            },
-            "stations": stations_json,
-            "items": items_json,
-            "players": [{"name": "robot", "x": 0, "y": 0, "direction": [0, 1]}],
-            "goal_description": "Make a cheese burger with cheese on top of the patty",
-            "goal": [
-                {"predicate": "iscooked", "args": ["patty"], "ids": [1]},
-                {"predicate": "atop", "args": ["topbun", "cheese"], "ids": [2, 3]},
-                {"predicate": "atop", "args": ["cheese", "patty"], "ids": [3, 1]},
-                {"predicate": "atop", "args": ["patty", "bottombun"], "ids": [1, 4]},
-            ],
-        }
-        return level_json
-
-
-class NoStationAtLocationError(Exception):
-    """Raised when trying to place an item at a location without a station."""
-
-    pass
-
-
-import os
 
 
 def render_level(level: LevelState, tile_size: int) -> pygame.Surface:
@@ -369,7 +199,7 @@ def main():
 
             button_y += button_height + 10
         except FileNotFoundError:
-            print(f"Asset not found: {asset_path}")
+            print(f"Asset not found: {item_asset_path}")
 
     run_file_button = pygame_gui.elements.UIButton(
         relative_rect=pygame.Rect((SCREEN_WIDTH, 816), (100, 50)),
